requirements.txt
Django==1.6
South==0.8.4
django-registration==1.0
django-social-auth==0.7.28
django-any==0.2.0
coverage==3.6